refactor: extract prediction utilities to break circular dependency
Move getPredictions, getMedia, insertPrediction, insertVideoPredictions,
aggregateDeploymentMetadata, and processMediaDeployment (plus their
private helpers) from importer.js into a new prediction.js module.

This breaks the circular import chain:
inference-consumer → importer → queue-scheduler → inference-consumer
