+++ src/main/services/import/index.js
@@ -2,5 +2,5 @@
  * Import services re-exports
  */
 
-export { Importer, getPredictions } from './importer.js'
+export { Importer } from './importer.js'
 export * from './parsers/index.js'
